refactor peer/fork indexes + better tests for underlying types
